refactor: separate geometry backend implementations
